Finish switch to SQL

+++ src/main/java/de/machag/HouseService/RESTController/HouseController.java
@@ -1,16 +1,15 @@
 package de.machag.HouseService.RESTController;
 
 import de.machag.HouseService.House.House;
-import de.machag.HouseService.House.HouseStatus;
 import de.machag.HouseService.House.HouseRepository;
-import de.machag.HouseService.Util;
+import de.machag.HouseService.House.HouseStatus;
+import org.apache.commons.collections4.IterableUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
-import java.util.UUID;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/house")
@@ -19,12 +18,9 @@ public class HouseController {
     @Autowired
     HouseRepository houseRepository;
 
-    @Autowired
-    Util utilities;
-
     @GetMapping("info/id={houseId}")
     @ResponseBody
-    public ResponseEntity getHouseById(@PathVariable String houseId) {
+    public ResponseEntity getHouseById(@PathVariable int houseId) {
         if(houseRepository.findById(houseId).isPresent() && houseRepository.findById(houseId).get() != null) {
             House targetHouse = houseRepository.findById(houseId).get();
             return ResponseEntity.ok(targetHouse.toString());
@@ -35,8 +31,9 @@ public class HouseController {
     @GetMapping("info/address={addressFull}")
     @ResponseBody
     public ResponseEntity getHouseByAdressFull(@PathVariable String addressFull) {
-        if(houseRepository.findByAddressFull(addressFull) != null) {
-            House targetHouse = houseRepository.findById(addressFull).get();
+        Optional<House> optionalHouse = houseRepository.findByAddressFullIgnoreCase(addressFull);
+        if(optionalHouse.isPresent()) {
+            House targetHouse = optionalHouse.get();
             return ResponseEntity.ok(targetHouse.toString());
         }
         return new ResponseEntity(HttpStatus.NOT_FOUND);
@@ -45,45 +42,38 @@ public class HouseController {
     @GetMapping("info/all")
     @ResponseBody
     public ResponseEntity getAllHouses() {
-        List<House> allHouses = houseRepository.findAll();
-        if(allHouses.size() > 0) {
-            return ResponseEntity.ok(utilities.fancyPrintAll(allHouses));
+        Iterable<House> allHouses = houseRepository.findAll();
+        if(IterableUtils.size(allHouses) > 0) {
+            return ResponseEntity.ok(allHouses.toString());
         }
         return new ResponseEntity(HttpStatus.NOT_FOUND);
     }
 
-    @PostMapping("create/{addressFull},{purchasePriceString},{rentPerMonthString},{statusString}")
+    @PostMapping("{addressFull},{purchasePrice},{rentPerMonth},{statusString}")
     @ResponseBody
-    public ResponseEntity createAccount(@PathVariable String addressFull,
-                                        @PathVariable String purchasePriceString,
-                                        @PathVariable String rentPerMonthString,
+    public ResponseEntity createHouse(@PathVariable String addressFull,
+                                        @PathVariable double purchasePrice,
+                                        @PathVariable double rentPerMonth,
                                         @PathVariable String statusString) {
 
-        double purchasePrice;
-        double rentPerMonth;
         HouseStatus houseStatus;
 
         try {
-            purchasePrice = Double.parseDouble(purchasePriceString);
-            rentPerMonth = Double.parseDouble(rentPerMonthString);
             houseStatus = HouseStatus.parseString(statusString);
-        } catch (NumberFormatException numberFormatException) {
-            return new ResponseEntity(HttpStatus.BAD_REQUEST);
         } catch (IllegalArgumentException illegalArgumentException) {
             return new ResponseEntity(HttpStatus.BAD_REQUEST);
         }
 
-        if(houseRepository.findByAddressFull(addressFull) != null) {
+        if(houseRepository.findByAddressFullIgnoreCase(addressFull).isPresent()) {
             return new ResponseEntity(HttpStatus.CONFLICT);
         }
 
         House newHouse = new House(
-                UUID.randomUUID().toString(),
                 addressFull,
                 purchasePrice,
                 rentPerMonth,
                 houseStatus,
-                null
+                0
         );
 
         houseRepository.save(newHouse);
@@ -94,10 +84,13 @@ public class HouseController {
     @DeleteMapping("{addressFull}")
     @ResponseBody
     public ResponseEntity deleteHouseByAddressFull(@PathVariable String addressFull) {
-        House targetHouse = houseRepository.findByAddressFull(addressFull);
-        if(targetHouse != null) {
-            houseRepository.delete(targetHouse);
-            return ResponseEntity.ok("Removed house: \n" + targetHouse.toString());
+        Optional<House> optionalHouse = houseRepository.findByAddressFullIgnoreCase(addressFull);
+        if(optionalHouse.isPresent()) {
+            House targetHouse = optionalHouse.get();
+            if(targetHouse != null) {
+                houseRepository.delete(targetHouse);
+                return ResponseEntity.ok("Removed house: \n" + targetHouse.toString());
+            }
         }
         return new ResponseEntity(HttpStatus.NOT_FOUND);
     }
